Move both of the classes (PhysicalPlanWithCost and PhysicalPlanWithCostUtils) into the utils package: se.liu.ida.hefquin.engine.queryproc.impl.optimizer.utils
